Anorm refactoring
